+++ example/lib/my_icons.dart
@@ -27,14 +27,14 @@ import 'package:flutter/widgets.dart';
 ///         - asset: fonts/my_icons.otf
 /// ```
 abstract final class MyIcons {
-  static const iconFontFamily = 'My Icons';
+  static const fontFamily = 'My Icons';
 
   /// arrow_right
   ///
   /// ![arrow_right](data:image/svg+xml,%3Csvg%20xmlns='http://www.w3.org/2000/svg'%20width='32'%20height='32'%20viewBox='0%200%2024%2024'%20fill='none'%3E%3Cpath%20d='M5%2012h14M12%205l7%207-7%207'%20stroke='grey'%20stroke-width='2'%20stroke-linecap='round'%20stroke-linejoin='round'/%3E%3C/svg%3E)
   static const IconData arrowRight = IconData(
     0xe000,
-    fontFamily: iconFontFamily,
+    fontFamily: fontFamily,
   );
 
   /// check
@@ -42,7 +42,7 @@ abstract final class MyIcons {
   /// ![check](data:image/svg+xml,%3Csvg%20xmlns='http://www.w3.org/2000/svg'%20width='32'%20height='32'%20viewBox='0%200%2024%2024'%20fill='none'%3E%3Cpath%20d='M20%206L9%2017l-5-5'%20stroke='grey'%20stroke-width='2'%20stroke-linecap='round'%20stroke-linejoin='round'/%3E%3C/svg%3E)
   static const IconData check = IconData(
     0xe001,
-    fontFamily: iconFontFamily,
+    fontFamily: fontFamily,
   );
 
   /// menu
@@ -50,7 +50,7 @@ abstract final class MyIcons {
   /// ![menu](data:image/svg+xml,%3Csvg%20xmlns='http://www.w3.org/2000/svg'%20width='32'%20height='32'%20viewBox='0%200%2024%2024'%20fill='none'%3E%3Cpath%20d='M4%207h16M4%2012h16M4%2017h16'%20stroke='grey'%20stroke-width='2'%20stroke-linecap='round'/%3E%3C/svg%3E)
   static const IconData menu = IconData(
     0xe002,
-    fontFamily: iconFontFamily,
+    fontFamily: fontFamily,
   );
 
   /// plus
@@ -58,6 +58,6 @@ abstract final class MyIcons {
   /// ![plus](data:image/svg+xml,%3Csvg%20xmlns='http://www.w3.org/2000/svg'%20width='32'%20height='32'%20viewBox='0%200%2024%2024'%20fill='none'%3E%3Cpath%20d='M12%205v14M5%2012h14'%20stroke='grey'%20stroke-width='2'%20stroke-linecap='round'/%3E%3C/svg%3E)
   static const IconData plus = IconData(
     0xe003,
-    fontFamily: iconFontFamily,
+    fontFamily: fontFamily,
   );
 }